# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "ncov19.us API",
        "description": "API Support: ncov19us@gmail.com | URL: https://github.com/ncov19-us/back-end | [GNU GENERAL PUBLIC LICENSE](https://github.com/ncov19-us/back-end/blob/master/LICENSE)",
        "version": "0.1.5"
    },
    "paths": {
...
        "/news": {
            "get": {
                "summary": "Get Gnews",
                "description": "Fetch US news from Google News API and return the results in JSON\n    ",
                "operationId": "get_gnews_news_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Post Gnews",
                "description": "Fetch specific state and topic news from Google News API and return the\nresults in JSON\n\nInput: NewsInput object schema, with state and topic attribute string\nOutput: JSONResponse of the topics fetched",
                "operationId": "post_gnews_news_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewsInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/twitter": {
            "get": {
                "summary": "Get Twitter",
                "description": "Fetch and return Twitter data from MongoDB connection.\n\n:param: none\n:return: str",
                "operationId": "get_twitter_twitter_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TwitterOutput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Post Twitter",
                "description": "Fetch and return Twitter data from MongoDB connection.\n\n:param: none. Two letter state abbreviation.\n:return: str",
                "operationId": "post_twitter_twitter_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TwitterInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TwitterOutput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/county": {
            "get": {
                "summary": "Get County Data",
                "description": "Get all US county data and return it as a big fat json string. Respond with\n404 if run into error.\n- Retrieves county locations, cached for 1 hour.\n\n:param: none.\n:return: JSONResponse",
                "operationId": "get_county_data_county_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountyOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Post County",
                "description": "Get all US county data and return it as a big fat json string. Respond with\n404 if run into error.\n- Retrieves county locations, cached for 1 hour.\n\n:param: none.\n:return: JSONResponse",
                "operationId": "post_county_county_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CountyInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountyOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/state": {
            "post": {
                "summary": "Post State",
                "description": "Fetch state level data time series for a single state, ignoring the \nunattributed and out of state cases.\n\nInput: two letter states code",
                "operationId": "post_state_state_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StateInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StateOutput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/country": {
            "post": {
                "summary": "Get Country",
                "description": "Fetch country level data time series for a single country\n\nInput: Two letter country alpha2Code",
                "operationId": "get_country_country_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CountryInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountryOutput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "summary": "Get Stats",
                "description": "Get overall tested, confirmed, and deaths stats from the database\nand return it as a json string. For the top bar.\n\n:param: none.\n:return: JSONResponse",
                "operationId": "get_stats_stats_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatsOutput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Post Stats",
                "description": "Get overall tested, confirmed, and deaths stats from the database\nand return it as a json string. For the top bar.\n\n:param: Stats\n:return: JSONResponse",
                "operationId": "post_stats_stats_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatsInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatsOutput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CountryInput": {
                "title": "CountryInput",
                "required": [
                    "alpha2Code"
                ],
                "type": "object",
                "properties": {
                    "alpha2Code": {
                        "title": "Alpha2Code",
                        "type": "string"
                    }
                }
            },
...
            "CountyInput": {
                "title": "CountyInput",
                "type": "object",
                "properties": {
                    "state": {
                        "title": "State",
                        "type": "string",
                        "default": "CA"
                    },
                    "county": {
                        "title": "County",
                        "type": "string",
                        "default": "Orange"
                    }
                }
            },
            "CountyOut": {
                "title": "CountyOut",
                "required": [
                    "success",
                    "message"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "title": "Success",
                        "type": "boolean"
                    },
                    "message": {
                        "title": "Message",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/County"
                        }
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "Message": {
                "title": "Message",
                "required": [
                    "message"
                ],
                "type": "object",
                "properties": {
                    "message": {
                        "title": "Message",
                        "type": "string"
                    }
                }
            },
...
            "NewsInput": {
                "title": "NewsInput",
                "type": "object",
                "properties": {
                    "state": {
                        "title": "State",
                        "type": "string",
                        "default": "CA"
                    },
                    "topic": {
                        "title": "Topic",
                        "type": "string",
                        "default": "Coronavirus"
                    }
                }
            },
            "NewsOut": {
                "title": "NewsOut",
                "required": [
                    "success",
                    "message"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "title": "Success",
                        "type": "boolean"
                    },
                    "message": {
                        "title": "Message",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/News"
                        }
                    }
                }
            },
...
            "StateInput": {
                "title": "StateInput",
                "required": [
                    "stateAbbr"
                ],
                "type": "object",
                "properties": {
                    "stateAbbr": {
                        "title": "Stateabbr",
                        "type": "string"
                    }
                }
            },
            "StateOutput": {
                "title": "StateOutput",
                "required": [
                    "success",
                    "message"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "title": "Success",
                        "type": "boolean"
                    },
                    "message": {
                        "title": "Message",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/State"
                        }
                    }
                }
            },
...
            "StatsOutput": {
                "title": "StatsOutput",
                "required": [
                    "success",
                    "message"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "title": "Success",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/Stats"
                    }
                }
            },
...
            "TwitterInput": {
                "title": "TwitterInput",
                "type": "object",
                "properties": {
                    "state": {
                        "title": "State",
                        "type": "string",
                        "default": "CA"
                    }
                }
            },
            "TwitterOutput": {
                "title": "TwitterOutput",
                "required": [
                    "success",
                    "message"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "title": "Success",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/UserTweets"
                    }
                }
            },
...
        }
    }
}